+++ Src/Private/Get-AbrVbrCloudConnectGP.ps1
@@ -6,7 +6,7 @@ function Get-AbrVbrCloudConnectGP {
     .DESCRIPTION
         Documents the configuration of Veeam VBR in Word/HTML/Text formats using PScribo.
     .NOTES
-        Version:        0.8.12
+        Version:        0.8.13
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -50,7 +50,7 @@ function Get-AbrVbrCloudConnectGP {
                             }
 
                             if ($HealthCheck.Jobs.BestPractice) {
-                                $OutObj | Where-Object { $Null -like $_.'Description' } | Set-Style -Style Warning -Property 'Description'
+                                $OutObj | Where-Object { $_.'Description' -eq "--" } | Set-Style -Style Warning -Property 'Description'
                                 $OutObj | Where-Object { $_.'Description' -match "Created by" } | Set-Style -Style Warning -Property 'Description'
                             }
 
@@ -65,8 +65,7 @@ function Get-AbrVbrCloudConnectGP {
                             }
                             $OutObj | Sort-Object -Property 'Name' | Table @TableParams
                             if ($HealthCheck.Jobs.BestPractice) {
-                                if ($OutObj | Where-Object { $_.'Description' -match 'Created by' -or $Null -like $_.'Description' }) {
-                                    Paragraph "Health Check:" -Bold -Underline
+                                if ($OutObj | Where-Object { $_.'Description' -match 'Created by' -or $_.'Description' -eq '--' }) {                                    Paragraph "Health Check:" -Bold -Underline
                                     BlankLine
                                     Paragraph {
                                         Text "Best Practice:" -Bold